backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
groq
mcp
firebase-admin
pydantic
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
